Trait rendy_graph::NodeDesc

pub trait NodeDesc<B: Backend, T: ?Sized>: Debug + Sized + 'static {
    type Node: Node<B, T>;
    fn build<'a>(
        self, 
        ctx: &GraphContext<B>, 
        factory: &mut Factory<B>, 
        family: &mut Family<B>, 
        queue: usize, 
        aux: &T, 
        buffers: Vec<NodeBuffer>, 
        images: Vec<NodeImage>
    ) -> Result<Self::Node, NodeBuildError>;

    fn builder(self) -> DescBuilder<B, T, Self> { ... }

    fn buffers(&self) -> Vec<BufferAccess> { ... }

    fn images(&self) -> Vec<ImageAccess> { ... }
}

Description of the node. Implementation of the builder type provide framegraph with static information about node that is used for building the node.

Associated Types

type Node: Node<B, T>

Node this builder builds.

Loading content...
Required methods

fn build<'a>(
self,
ctx: &GraphContext<B>,
factory: &mut Factory<B>,
family: &mut Family<B>,
queue: usize,
aux: &T,
buffers: Vec<NodeBuffer>,
images: Vec<NodeImage>
) -> Result<Self::Node, NodeBuildError>

Build the node.

Parameters

factory - factory instance. aux - auxiliary data. family - id of the family this node will be executed on. resources - set of transient resources managed by graph. with barriers required for interface resources.

Loading content...
Provided methods

fn builder(self) -> DescBuilder<B, T, Self>

Make node builder.

fn buffers(&self) -> Vec<BufferAccess>

Get set or buffer resources the node uses.

fn images(&self) -> Vec<ImageAccess>

Get set or image resources the node uses.

Loading content...